# [python] share census objects in tests for faster test time