--- a/api/python/cellxgene_census/tests/conftest.py
+++ b/api/python/cellxgene_census/tests/conftest.py
@@ -1,6 +1,7 @@
 import pytest
 import tiledbsoma as soma
 
+import cellxgene_census
 from cellxgene_census import get_default_soma_context
 
 TEST_MARKERS_SKIPPED_BY_DEFAULT = ["expensive", "experimental"]
@@ -49,3 +50,16 @@ def pytest_configure(config: pytest.Config) -> None:
 def small_mem_context() -> soma.SOMATileDBContext:
     """used to keep memory usage smaller for GHA runners."""
     return get_default_soma_context(tiledb_config={"soma.init_buffer_bytes": 32 * 1024**2})
+
+
+# Fixtures for census objects
+
+
+@pytest.fixture(scope="session")
+def census() -> soma.Collection:
+    return cellxgene_census.open_soma(census_version="latest")
+
+
+@pytest.fixture(scope="session")
+def lts_census() -> soma.Collection:
+    return cellxgene_census.open_soma(census_version="stable")