I propose we re-use SOMA objects in testing to cut down on total test time.

From a single measurement, I see a little more than a 10% (~40 seconds) decrease in test time with this change. I've run the `get_anndata` tests on this commit and the one before it. I haven't run this many times, so I don't know what the variance is. You can see that with this commit there is far lower time taken on setup for each task (since the resource is simply removed)

### Timings on this commit

<details>

```
================================================================================ slowest durations ================================================================================
53.89s call     tests/test_get_anndata.py::test_get_anndata_allows_missing_obs_or_var_filter
34.26s call     tests/test_get_anndata.py::test_get_anndata_value_filter
33.00s call     tests/test_get_anndata.py::test_get_anndata_obs_embeddings[obs_embeddings0]
31.75s call     tests/test_get_anndata.py::test_get_anndata_two_layers[layers0]
30.18s call     tests/test_get_anndata.py::test_get_anndata_two_layers[layers1]
16.68s call     tests/test_get_anndata.py::test_get_anndata_var_embeddings[var_embeddings0]
13.84s call     tests/test_get_anndata.py::test_get_anndata_x_layer[raw]
13.23s call     tests/test_get_anndata.py::test_get_anndata_obsm_one_layer[geneformer]
12.33s call     tests/test_get_anndata.py::test_get_anndata_obsm_one_layer[scvi]
7.79s call     tests/test_get_anndata.py::test_get_anndata_x_layer[normalized]
7.64s call     tests/test_get_anndata.py::test_get_anndata_obsm_two_layers[obsm_layers0]
6.23s call     tests/test_get_anndata.py::test_get_anndata_coords
6.18s setup    tests/test_get_anndata.py::test_get_anndata_value_filter
1.29s setup    tests/test_get_anndata.py::test_get_anndata_obsm_one_layer[scvi]
0.00s teardown tests/test_get_anndata.py::test_get_anndata_obsm_layers_and_add_obs_embedding_fails
0.00s call     tests/test_get_anndata.py::test_get_anndata_obsm_layers_and_add_obs_embedding_fails
0.00s setup    tests/test_get_anndata.py::test_get_anndata_obsm_layers_and_add_obs_embedding_fails
0.00s teardown tests/test_get_anndata.py::test_get_anndata_var_embeddings[var_embeddings0]
0.00s call     tests/test_get_anndata.py::test_get_anndata_wrong_layer_names
0.00s setup    tests/test_get_anndata.py::test_get_anndata_x_layer[normalized]
0.00s setup    tests/test_get_anndata.py::test_get_anndata_x_layer[raw]
0.00s setup    tests/test_get_anndata.py::test_get_anndata_two_layers[layers1]
0.00s setup    tests/test_get_anndata.py::test_get_anndata_two_layers[layers0]
0.00s setup    tests/test_get_anndata.py::test_get_anndata_obsm_two_layers[obsm_layers0]
0.00s teardown tests/test_get_anndata.py::test_get_anndata_obs_embeddings[obs_embeddings0]
0.00s setup    tests/test_get_anndata.py::test_get_anndata_obs_embeddings[obs_embeddings0]
0.00s setup    tests/test_get_anndata.py::test_get_anndata_var_embeddings[var_embeddings0]
0.00s setup    tests/test_get_anndata.py::test_get_anndata_obsm_one_layer[geneformer]
0.00s teardown tests/test_get_anndata.py::test_get_anndata_obsm_one_layer[geneformer]
0.00s teardown tests/test_get_anndata.py::test_get_anndata_two_layers[layers0]
0.00s teardown tests/test_get_anndata.py::test_get_anndata_x_layer[normalized]
0.00s teardown tests/test_get_anndata.py::test_get_anndata_x_layer[raw]
0.00s teardown tests/test_get_anndata.py::test_get_anndata_obsm_two_layers[obsm_layers0]
0.00s teardown tests/test_get_anndata.py::test_get_anndata_two_layers[layers1]
0.00s setup    tests/test_get_anndata.py::test_get_anndata_coords
0.00s teardown tests/test_get_anndata.py::test_get_anndata_obsm_one_layer[scvi]
0.00s setup    tests/test_get_anndata.py::test_get_anndata_wrong_layer_names
0.00s teardown tests/test_get_anndata.py::test_get_anndata_value_filter
0.00s teardown tests/test_get_anndata.py::test_get_anndata_allows_missing_obs_or_var_filter
0.00s setup    tests/test_get_anndata.py::test_get_anndata_allows_missing_obs_or_var_filter
0.00s teardown tests/test_get_anndata.py::test_get_anndata_coords
0.00s teardown tests/test_get_anndata.py::test_get_anndata_wrong_layer_names
=================================================================== 14 passed, 34 warnings in 268.35s (0:04:28) ===================================================================
```

</details>

### Timings on main

<details>

```
================================================================================ slowest durations ================================================================================
44.75s call     tests/test_get_anndata.py::test_get_anndata_allows_missing_obs_or_var_filter
35.88s call     tests/test_get_anndata.py::test_get_anndata_obs_embeddings[obs_embeddings0]
34.50s call     tests/test_get_anndata.py::test_get_anndata_two_layers[layers1]
33.52s call     tests/test_get_anndata.py::test_get_anndata_two_layers[layers0]
28.05s call     tests/test_get_anndata.py::test_get_anndata_value_filter
19.20s call     tests/test_get_anndata.py::test_get_anndata_var_embeddings[var_embeddings0]
15.72s call     tests/test_get_anndata.py::test_get_anndata_obsm_two_layers[obsm_layers0]
13.25s call     tests/test_get_anndata.py::test_get_anndata_obsm_one_layer[scvi]
12.79s call     tests/test_get_anndata.py::test_get_anndata_obsm_one_layer[geneformer]
11.51s call     tests/test_get_anndata.py::test_get_anndata_x_layer[raw]
11.35s call     tests/test_get_anndata.py::test_get_anndata_x_layer[normalized]
9.00s call     tests/test_get_anndata.py::test_get_anndata_coords
6.48s setup    tests/test_get_anndata.py::test_get_anndata_value_filter
5.26s setup    tests/test_get_anndata.py::test_get_anndata_coords
5.18s setup    tests/test_get_anndata.py::test_get_anndata_two_layers[layers0]
5.18s setup    tests/test_get_anndata.py::test_get_anndata_obsm_one_layer[geneformer]
3.15s setup    tests/test_get_anndata.py::test_get_anndata_x_layer[raw]
3.06s setup    tests/test_get_anndata.py::test_get_anndata_two_layers[layers1]
2.71s call     tests/test_get_anndata.py::test_get_anndata_wrong_layer_names
1.38s setup    tests/test_get_anndata.py::test_get_anndata_x_layer[normalized]
1.31s setup    tests/test_get_anndata.py::test_get_anndata_allows_missing_obs_or_var_filter
1.29s setup    tests/test_get_anndata.py::test_get_anndata_obsm_two_layers[obsm_layers0]
1.19s setup    tests/test_get_anndata.py::test_get_anndata_obsm_layers_and_add_obs_embedding_fails
1.15s setup    tests/test_get_anndata.py::test_get_anndata_obsm_one_layer[scvi]
1.14s setup    tests/test_get_anndata.py::test_get_anndata_obs_embeddings[obs_embeddings0]
1.11s setup    tests/test_get_anndata.py::test_get_anndata_wrong_layer_names
1.08s setup    tests/test_get_anndata.py::test_get_anndata_var_embeddings[var_embeddings0]
0.65s call     tests/test_get_anndata.py::test_get_anndata_obsm_layers_and_add_obs_embedding_fails
0.00s teardown tests/test_get_anndata.py::test_get_anndata_var_embeddings[var_embeddings0]
0.00s teardown tests/test_get_anndata.py::test_get_anndata_wrong_layer_names
0.00s teardown tests/test_get_anndata.py::test_get_anndata_obsm_layers_and_add_obs_embedding_fails
0.00s teardown tests/test_get_anndata.py::test_get_anndata_obs_embeddings[obs_embeddings0]
0.00s teardown tests/test_get_anndata.py::test_get_anndata_obsm_two_layers[obsm_layers0]
0.00s teardown tests/test_get_anndata.py::test_get_anndata_obsm_one_layer[geneformer]
0.00s teardown tests/test_get_anndata.py::test_get_anndata_value_filter
0.00s teardown tests/test_get_anndata.py::test_get_anndata_x_layer[raw]
0.00s teardown tests/test_get_anndata.py::test_get_anndata_x_layer[normalized]
0.00s teardown tests/test_get_anndata.py::test_get_anndata_two_layers[layers0]
0.00s teardown tests/test_get_anndata.py::test_get_anndata_two_layers[layers1]
0.00s teardown tests/test_get_anndata.py::test_get_anndata_coords
0.00s teardown tests/test_get_anndata.py::test_get_anndata_allows_missing_obs_or_var_filter
0.00s teardown tests/test_get_anndata.py::test_get_anndata_obsm_one_layer[scvi]
=================================================================== 14 passed, 34 warnings in 310.89s (0:05:10) ===================================================================
```

</details>