@prathapsridharan 2024-05-22
If all you intend to do is change the `scope` of the `census` `pytest.fixture` from the default of `function` to `session`, then I think it is reasonable to make that change in a single PR in all the tests. But I would also like @ebezzi opinion here too