@ivirshup 2024-05-22
Cool. So the two other sets of tests where I think this can be used is `test_lts_compat` and `test_get_helpers.py`. I would think we should also share the fixture between the tests, so can move it to the conftest.